cli: Extract command runtime parsing and thin app.rs dispatcher
Move parse-time conversion, clap error handling, help rendering bridges, and
convert_* helpers into services/parse/command_runtime.rs. Extract output
rendering and execution-phase logging into services/app_support.rs to keep
app.rs focused on startup lifecycle and thin parse/execute/render orchestration.
Register version and completion default constructors so CommandRegistry covers
the full current command catalog.
