+++ context/architecture.md
@@ -90,9 +90,9 @@ The repository includes a new placeholder Rust binary crate at `cli/`.
 - `cli/src/main.rs` is the executable entrypoint (`sce`) and delegates to `app::run`.
 - `cli/src/cli_schema.rs` defines the clap-based CLI schema using derive macros for all top-level commands and subcommands, and renders command-local help text for the `auth` command tree (`auth`, `auth login`, `auth logout`, `auth status`).
 - `cli/src/app.rs` provides the clap-based argument dispatch loop with deterministic help/setup execution, bare-command help routing for `sce auth` and `sce config`, centralized stream routing (`stdout` success payloads, `stderr` redacted diagnostics), stable class-based exit-code mapping (`2` parse, `3` validation, `4` runtime, `5` dependency), and stable class-based stderr diagnostic codes (`SCE-ERR-PARSE`, `SCE-ERR-VALIDATION`, `SCE-ERR-RUNTIME`, `SCE-ERR-DEPENDENCY`) with default `Try:` remediation injection when missing.
-- The app runtime now moves through explicit startup phases in `cli/src/app.rs`: dependency bootstrapping (`perform_dependency_check`), startup context construction (`build_startup_context`), runtime initialization (`initialize_runtime`), command parse/execute inside telemetry subscriber context (`run_command_lifecycle`, `parse_command_phase`, `execute_command_phase`), and final output rendering (`render_run_outcome`). `AppRuntime` carries a minimal `AppContext` plus startup-diagnostic state across those phases.
+- The app runtime now moves through explicit startup phases in `cli/src/app.rs`: dependency bootstrapping (`perform_dependency_check`), startup context construction (`build_startup_context`), runtime initialization (`initialize_runtime`), command parse/execute inside telemetry subscriber context (`run_command_lifecycle`, `parse_command_phase` plus `services::app_support::execute_command_phase`), and final output rendering through `services::app_support::render_run_outcome`. `AppRuntime` carries a minimal `AppContext`, static `CommandRegistry`, and startup-diagnostic state across those phases.
 - `AppContext` is the CLI dependency-injection container in `cli/src/app.rs`: it holds `Arc<dyn Logger>`, `Arc<dyn Telemetry>`, `Arc<dyn FsOps>`, and `Arc<dyn GitOps>`, constructed once during runtime initialization from the production logger, telemetry runtime, filesystem ops, and git ops implementations.
-- Command parse-time conversion and run-time handling are separated by an internal `RuntimeCommand` seam. The `RuntimeCommand` trait and `RuntimeCommandHandle` type alias live in `cli/src/services/command_registry.rs`, alongside the `CommandRegistry` struct that maps command names to zero-arg constructor functions. `parse_command_phase` in `cli/src/app.rs` delegates clap-output conversion to the nested `command_runtime` module, which returns boxed command objects that own their own `name()` and `execute(&AppContext)` behavior. Service-owned `command.rs` modules define the migrated command structs for help/version/completion/auth/config/setup/doctor/hooks. `execute_command_phase` emits lifecycle logs around `command.execute(...)` instead of switching over one central app-level dispatch `match`.
+- Command parse-time conversion and run-time handling are separated by an internal `RuntimeCommand` seam. The `RuntimeCommand` trait and `RuntimeCommandHandle` type alias live in `cli/src/services/command_registry.rs`, alongside the `CommandRegistry` struct that maps command names to zero-arg constructor functions. `parse_command_phase` in `cli/src/app.rs` delegates clap-output conversion to `cli/src/services/parse/command_runtime.rs`, which owns clap error classification, help rendering bridges, and parsed-request-to-command conversion while returning boxed command objects with `name()` and `execute(&AppContext)` behavior. Service-owned `command.rs` modules define the migrated command structs for help/version/completion/auth/config/setup/doctor/hooks. `services::app_support::execute_command_phase` emits lifecycle logs around `command.execute(...)` instead of switching over one central app-level dispatch `match`.
 - Startup observability bootstrapping in `cli/src/app.rs` still tolerates invalid default-discovered config files by continuing with degraded defaults plus `sce.config.invalid_config` warn-level logs, but the warning/logging work is now isolated behind the startup-context and runtime-initialization phases rather than one inline startup function.
 - `cli/src/services/observability.rs` provides deterministic runtime observability controls and rendering for app lifecycle logs, including shared config-resolved threshold/format and OTEL/file-sink inputs with precedence `env > config file > defaults` for non-flag observability keys, optional file sink controls (`SCE_LOG_FILE`, `SCE_LOG_FILE_MODE` with deterministic truncate-or-append policy), optional OTEL export bootstrap (`SCE_OTEL_ENABLED`, `OTEL_EXPORTER_OTLP_ENDPOINT`, `OTEL_EXPORTER_OTLP_PROTOCOL`), stable event identifiers, severity filtering, the forced-emission warning path used for invalid discovered config startup diagnostics, stderr-only primary emission with optional mirrored file writes, and redaction-safe emission through the shared security helper. Its `observability::traits` submodule exposes the current `Logger` and object-safe `Telemetry` trait boundaries plus `NoopLogger`; the concrete observability logger and telemetry runtime still own behavior and implement those traits without changing runtime behavior.
 - `cli/src/services/observability.rs` no longer owns duplicate log/OTEL enums or parsing helpers; it consumes the canonical primitive seam from `cli/src/services/config/mod.rs` and stays focused on logger and telemetry runtime behavior.
