+++ context/sce/cli-shared-output-format-contract.md
@@ -11,7 +11,7 @@
 ## Current command integration
 
 - `cli/src/services/config/mod.rs` uses the shared type through `ReportFormat = OutputFormat`.
-- `cli/src/services/version.rs` uses the shared type through `VersionFormat = OutputFormat`.
+- `cli/src/services/version/mod.rs` uses the shared type through `VersionFormat = OutputFormat`.
 - Both commands keep deterministic default `text` when `--format` is omitted.
 - Both commands keep stable `--format <text|json>` usage in help text and parser behavior.
